refactor(generator): split DataRecords out of DiscoveryData, trim top file to 135 lines
